add phase and refactor
