+++ .eslintignore
@@ -4,4 +4,5 @@
 **/esm/*
 
 # Test fixtures
+**/fixture/*
 **/__fixtures__/*
